finished refactor of proposalDetails
